style: adjust default style
